Update scripts for checking requirements and deploment

+++ local_server/dags/scripts/create_env.sh
@@ -1,22 +1,28 @@
 #!/bin/bash
 set -e
-if [ -d "$HOME/conda/envs/training_env" ]
+REPO_NAME=local_mlops
+LOCAL_REPO_DIRECTORY=$HOME
+REPO_OWNER=phandaiduonghcb
+
+cd $LOCAL_REPO_DIRECTORY/$REPO_NAME
+git fetch origin main
+output=$(git diff origin/main -- requirements-dev.txt)
+git pull
+if [[ -n "$output" ]]
 then
-    rm -rf "$HOME/conda/envs/training_env"
+    if [ -d "$HOME/conda/envs/training_env" ]
+    then
+        rm -rf "$HOME/conda/envs/training_env"
+    fi
+    conda create -n training_env
+    conda activate training_env
+    conda env list
+    conda install pip -n training_env
+    $HOME/conda/envs/training_env/bin/pip install -r $LOCAL_REPO_DIRECTORY/$REPO_NAME/requirements-dev.txt --no-cache-dir
 fi
 
-conda create -n training_env
-conda activate training_env
-conda env list
-conda install pip -n training_env
-cd $HOME/conda/envs/training_env
-git clone https://github.com/phandaiduonghcb/local_mlops
-# echo "Install dependencies..."
-$HOME/conda/envs/training_env/bin/pip install -r ./local_mlops/requirements-dev.txt --no-cache-dir
-$HOME/conda/envs/training_env/bin/pip cache purge
-pip cache purge
-cd local_mlops
 echo "Current working dir: $PWD"
+conda activate training_env
 dvc pull -AaR -r local-mlops-bucket
 ls ./ml/data/*
 echo "Create env successfully!"
